--- modules/convex_hull.py
@@ -1,28 +1,0 @@
-#!/usr/bin/env python3
-
-import importlib
-
-
-class ConvexHullPoint(tuple):
-    pass
-
-
-def multilayer_convex_hull(point_batch: list[ConvexHullPoint], layers: int = 1):
-    pyhull_convex_hull = importlib.import_module('pyhull.convex_hull').ConvexHull
-    hull_layers_points = []
-    self_hull_points = list(point_batch)
-    for layer in range(layers):
-        if len(self_hull_points) == 0:
-            break
-        hull = pyhull_convex_hull(self_hull_points)
-        hull_vertexes = set(vertex for hull_vertex in hull.vertices for vertex in hull_vertex)
-        hull_points = list(self_hull_points[vertex] for vertex in hull_vertexes)
-        if len(hull_points) > 0:
-            for hull_point in hull_points:
-                self_hull_points.remove(hull_point)
-                hull_layers_points.append(hull_point)
-        else:
-            hull_layers_points.extend(self_hull_points)
-    return hull_layers_points
-
-
